feat(codec): move codec
